Refactor: move widget type definitions to single namespace file
