+++ src/AreaChart/components/AreaChartContainer.ts
@@ -1,5 +1,7 @@
 import { SFC, createElement } from "react";
-import LineChartContainer, { LineChartContainerProps } from "../../LineChart/components/LineChartContainer";
+import LineChartContainer from "../../LineChart/components/LineChartContainer";
+import { Container } from "../../utils/namespaces";
+import LineChartContainerProps = Container.LineChartContainerProps;
 
 const AreaChartContainer: SFC<LineChartContainerProps> = props =>
     createElement(LineChartContainer, { ...props as LineChartContainerProps, fill: true });
