+++ src/BarChart/components/BarChartContainer.ts
@@ -1,20 +1,11 @@
-import { Component, ReactChild, createElement } from "react";
+import { Component, createElement } from "react";
 
 import { BarChart } from "./BarChart";
 import { ChartContainer } from "../../components/ChartContainer";
-import { SeriesData, SeriesProps, validateSeriesProps } from "../../utils/data";
-import { Dimensions } from "../../utils/style";
-import { BarLayoutProps, WrapperProps } from "../../utils/types";
-
-export interface BarChartContainerProps extends WrapperProps, Dimensions, BarLayoutProps {
-    series: SeriesProps[];
-}
-
-interface BarChartContainerState {
-    alertMessage?: ReactChild;
-    data?: SeriesData[];
-    loading?: boolean;
-}
+import { validateSeriesProps } from "../../utils/data";
+import { Container } from "../../utils/namespaces";
+import BarChartContainerProps = Container.BarChartContainerProps;
+import BarChartContainerState = Container.BarChartContainerState;
 
 export default class BarChartContainer extends Component<BarChartContainerProps, BarChartContainerState> {
     static defaultProps: Partial<BarChartContainerProps> = { orientation: "bar" };
