+++ src/utils/data.ts
@@ -1,51 +1,11 @@
 import { ReactChild, createElement } from "react";
 import { Datum } from "plotly.js";
-import { LineMode } from "./types";
-
-export interface DataSourceProps {
-    dataSourceMicroflow: string;
-    dataSourceType: "XPath" | "microflow";
-    entityConstraint: string;
-    dataEntity: string;
-}
-
-export interface SeriesDataSourceProps extends DataSourceProps {
-    xValueAttribute: string;
-    xValueSortAttribute: string;
-    sortOrder: SortOrder;
-    yValueAttribute: string;
-}
-
-export type MxObject = mendix.lib.MxObject;
-export type SortOrder = "asc" | "desc";
-
-export interface EventProps {
-    onClickEvent: "doNothing" | "showPage" | "callMicroflow";
-    onClickPage: string;
-    onClickMicroflow: string;
-    tooltipForm: string;
-}
-
-export interface SeriesProps extends SeriesDataSourceProps, EventProps {
-    name: string;
-    seriesOptions: string;
-}
-
-export interface LineSeriesProps extends SeriesProps {
-    mode?: LineMode;
-    lineColor: string;
-    lineStyle: "linear" | "spline";
-}
-
-export interface SeriesData {
-    data: mendix.lib.MxObject[];
-    series: SeriesProps;
-}
-
-export interface ScatterTrace {
-    x: Datum[];
-    y: number[] | Datum[];
-}
+import { Data } from "./namespaces";
+import SeriesProps = Data.SeriesProps;
+import SeriesData = Data.SeriesData;
+import SortOrder = Data.SortOrder;
+import EventProps = Data.EventProps;
+import ScatterTrace = Data.ScatterTrace;
 
 export const validateSeriesProps = <T extends Partial<SeriesProps>>(dataSeries: T[], widgetId: string, layoutOptions: string): ReactChild => { // tslint:disable-line max-line-length
     if (dataSeries && dataSeries.length) {
@@ -97,7 +57,7 @@ export const validateAdvancedOptions = (rawData: string): string => {
     return "";
 };
 
-export const fetchSeriesData = (mxObject: MxObject, series: SeriesProps): Promise<SeriesData> =>
+export const fetchSeriesData = (mxObject: mendix.lib.MxObject, series: SeriesProps): Promise<SeriesData> =>
     new Promise<SeriesData>((resolve, reject) => {
         if (series.dataEntity) {
             if (series.dataSourceType === "XPath") {
@@ -114,7 +74,7 @@ export const fetchSeriesData = (mxObject: MxObject, series: SeriesProps): Promis
         }
     });
 
-export const fetchByXPath = (guid: string, entity: string, constraint: string, sortBy?: string, sortOrder: SortOrder = "asc"): Promise<MxObject[]> =>
+export const fetchByXPath = (guid: string, entity: string, constraint: string, sortBy?: string, sortOrder: SortOrder = "asc"): Promise<mendix.lib.MxObject[]> =>
     new Promise((resolve, reject) => {
         const entityPath = entity.split("/");
         const entityName = entityPath.length > 1 ? entityPath[entityPath.length - 1] : entity;
@@ -129,17 +89,17 @@ export const fetchByXPath = (guid: string, entity: string, constraint: string, s
         });
     });
 
-export const fetchByMicroflow = (actionname: string, guid: string): Promise<MxObject[]> =>
+export const fetchByMicroflow = (actionname: string, guid: string): Promise<mendix.lib.MxObject[]> =>
     new Promise((resolve, reject) => {
         const errorMessage = `An error occurred while retrieving data by microflow (${actionname}): `;
         window.mx.ui.action(actionname, {
-            callback: (mxObjects: MxObject[]) => resolve(mxObjects),
+            callback: (mxObjects: mendix.lib.MxObject[]) => resolve(mxObjects),
             error: error => reject(`${errorMessage} ${error.message}`),
             params: { applyto: "selection", guids: [ guid ] }
         });
     });
 
-export const handleOnClick = <T extends EventProps>(options: T, mxObject?: MxObject) => {
+export const handleOnClick = <T extends EventProps>(options: T, mxObject?: mendix.lib.MxObject) => {
     if (!mxObject || options.onClickEvent === "doNothing") {
         return;
     }
